server/requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
grpc-tools==0.0.1
grpcio==1.27.2
grpcio-tools==1.27.2
gunicorn==20.0.4
numpy==1.18.1
opencv-python==4.2.0.32
protobuf==3.11.3
psutil==5.7.0
psycopg2-binary==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
